重构task-examples为chok-scheduler
