+++ oauth2-examples/chok-scheduler/pom.xml
@@ -5,11 +5,11 @@
 
 	<modelVersion>4.0.0</modelVersion>
 	<!--本Project識別 -->
-	<artifactId>task-examples</artifactId>
+	<artifactId>chok-scheduler</artifactId>
 	<!--表示打包Project的型態,可能為Jar、war、ear或pom，若是使用了android 則為apk -->
 	<packaging>war</packaging>
 	<!--以下是給工具看的,主要是本Project的資訊 -->
-	<name>task-examples</name>
+	<name>chok-scheduler</name>
 
 	<parent>
 		<groupId>com.chok</groupId>
@@ -180,7 +180,7 @@
 				</includes>
 			</resource>
 		</resources>
-		<finalName>task-examples</finalName>
+		<finalName>chok-scheduler</finalName>
 	</build>
 
 </project>
