+++ oauth2-examples/chok-scheduler/src/main/java/com/webconfig/DruidStatViewConfig.java
