+++ oauth2-examples/chok-scheduler/src/main/resources/application-dev.properties
