+++ oauth2-examples/chok-scheduler/src/main/resources/application-local-dev.properties
