+++ oauth2-examples/chok-scheduler/src/main/resources/application.properties
@@ -6,7 +6,7 @@ spring.profiles.active=@profileActive@
 # =====================================================================================
 # app-name
 # =====================================================================================
-spring.application.name=task-examples
+spring.application.name=chok-scheduler
 
 # =====================================================================================
 # i18n
