+++ oauth2-examples/chok-scheduler/src/main/resources/datasource-jdbc-prd.properties
