+++ oauth2-examples/chok-scheduler/src/main/resources/datasource-mybatis-local-uat.properties
