+++ oauth2-examples/chok-scheduler/src/main/resources/mybatis.xml
