+++ oauth2-examples/chok-scheduler/src/main/resources/redisson-single-local-uat.yml
