+++ oauth2-examples/chok-scheduler/src/main/resources/redisson-single-uat.yml
